pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ytflashback"
version = "1.0.0"
authors = [
    {name = "Akrm Al-Hakimi", email = "alhakimiakrmj@gmail.com"},
]
description = "A tool for searching YouTube videos by year, helping users find older content that might be buried by YouTube's algorithm"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Utilities",
]
keywords = ["youtube", "search", "video", "cli", "tui", "archive"]
dependencies = [
    "google-api-python-client",
    "python-dotenv",
    "click",
    "colorama", 
    "tabulate",
    "textual",
    "platformdirs",
]

[project.urls]
Homepage = "https://github.com/cachebag/flashback"
Repository = "https://github.com/cachebga/flashback.git"
Issues = "https://github.com/cachebag/flashback/issues"

[project.scripts]
ytflashback = "flashback.main:main_tui"
ytflashback-cli = "flashback.main:main_cli"

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/flashback"]

[tool.hatch.version]
path = "src/flashback/__init__.py" 